+++ common/src/main/java/dynamic_asset_generator/DynamicAssetGeneratorClient.java
@@ -1,26 +1,32 @@
 package dynamic_asset_generator;
 
+import dynamic_asset_generator.client.api.DynAssetGeneratorClientAPI;
+import dynamic_asset_generator.client.api.ForegroundTransferType;
+import dynamic_asset_generator.client.api.PaletteExtractor;
+import dynamic_asset_generator.client.util.IPalettePlan;
+import net.minecraft.resources.ResourceLocation;
+
 public class DynamicAssetGeneratorClient {
     public static void init() {
         //testing
-/*
-        String background = "textures/block/andesite.png";
+
+        String background = "textures/block/stone.png";
         PaletteExtractor extractor = new PaletteExtractor(new ResourceLocation("minecraft","textures/block/stone.png"),
-                new ResourceLocation("minecraft","textures/block/redstone_ore.png"),0,true, true);
+                new ResourceLocation("minecraft","textures/block/copper_ore.png"),6,true, true,0.2);
         IPalettePlan plan = new ForegroundTransferType(extractor, new ResourceLocation("minecraft",background),
                 true, false);
         PaletteExtractor extractor2 = new PaletteExtractor(new ResourceLocation("minecraft","textures/block/stone.png"),
-                new ResourceLocation("minecraft","textures/block/redstone_ore.png"), 0,true, false);
+                new ResourceLocation("minecraft","textures/block/redstone_ore.png"), 6,true, true,0.2);
         IPalettePlan plan2 = new ForegroundTransferType(extractor2, new ResourceLocation("minecraft",background),
                 true, false);
         PaletteExtractor extractor3 = new PaletteExtractor(new ResourceLocation("minecraft","textures/block/stone.png"),
-                new ResourceLocation("minecraft","textures/block/redstone_ore.png"), 0);
+                new ResourceLocation("minecraft","textures/block/coal_ore.png"), 6,true,true,0.2);
         IPalettePlan plan3 = new ForegroundTransferType(extractor3, new ResourceLocation("minecraft",background),
                 true, false);
         DynAssetGeneratorClientAPI.planPaletteCombinedImage(new ResourceLocation("minecraft","textures/block/end_stone.png"), plan);
         DynAssetGeneratorClientAPI.planPaletteCombinedImage(new ResourceLocation("minecraft","textures/block/cobblestone.png"), plan2);
         DynAssetGeneratorClientAPI.planPaletteCombinedImage(new ResourceLocation("minecraft","textures/block/tuff.png"), plan3);
-*/
+
         /*
         DynAssetGenClientPlanner.planPaletteCombinedImage(new ResourceLocation("minecraft","textures/item/gold_ingot.png"),
                 new PlannedPaletteCombinedImage(new ResourceLocation("minecraft","textures/block/moss_block.png"), new ResourceLocation("dynamic_asset_generator","textures/empty.png"), new ResourceLocation("minecraft","textures/item/copper_ingot.png"), false, 6, true));